+++ test-packages/custom-html-app/public/custom-index.html
@@ -4,25 +4,21 @@
 <head>
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title>Ignored Index.html</title>
+  <title>custom index</title>
   <meta name="description" content="">
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
-  {{content-for 'head'}}
+  <!-- EMBER_CLI_FASTBOOT_HEAD -->
 
   <link rel="stylesheet" href="assets/vendor.css">
   <link rel="stylesheet" href="assets/custom-html-file.css">
-
-  {{content-for 'head-footer'}}
 </head>
 
 <body>
-  {{content-for 'body'}}
+  <!-- EMBER_CLI_FASTBOOT_BODY -->
 
   <script src="assets/vendor.js"></script>
   <script src="assets/custom-html-file.js"></script>
-
-  {{content-for 'body-footer'}}
 </body>
 
 </html>
